+++ teensy3/WProgram.h
@@ -50,9 +50,9 @@ void tone(uint8_t pin, uint16_t frequency, uint32_t duration = 0);
 void noTone(uint8_t pin);
 
 // WMath prototypes
-uint32_t random(void);
-uint32_t random(uint32_t howbig);
-int32_t random(int32_t howsmall, int32_t howbig);
+long random(void);
+long random(uint32_t howbig);
+long random(int32_t howsmall, int32_t howbig);
 void randomSeed(uint32_t newseed);
 void srandom(uint32_t newseed);
 long map(long, long, long, long, long);
